requirements2.txt
langchain>=0.3.26
# langchain-comunity==0.3.27 # doesn't work: pip install langchain-community langchain-core
langchain-chroma==0.2.4
langchain-openai==0.3.27
numpy==1.26.4
pandas==2.3.0
matplotlib==3.10.1
seaborn==0.13.2
tensorboardX==2.6.4
cytoolz==1.0.1
fastparquet==2024.11.0
pyarrow==20.0.0
torch==2.7.1
pyyaml==6.0.2
streamlit==1.46.1
sentence-transformers==5.0.0
# make sure to get your OpenAI API key